Styled editors sections and added functionality
Styled and organized the editors section.

Added localStorage for data persistence.

Added buttons to for exporting and printing the CV
which still need to be implemented.

Added general CV configuration editor.

Sections are now indexed when created and sorted at the provider.

Started creating the draggable field component for dynamically
organizing sections and inputs.