Fix drag and drop and improve app
Moved the drag and drop state out of the drag container component
and put it in a context to allow more effective and context-aware
functionality. Sections and fields can now work independently and
do not respond to each other's drag events.

Deconstructed the DropDownContainer stylings in the EditorContainer
stylesheet to allow the drop down's content to expand the full width
of the drag handle and delete button. This was breaking the mobile view.

Refactored CVEditors component.

Updated lots of UI and styling components to accommodate drag and drop.

Various other bug fixes and update.