Implemented AddSection Fixed State Bug
After implementing AddSection a bug was revealed with how state was
being managed. The way I was calling the EditorContainer component as
a function was forcing the hooks to become out of sync and not following
React best practices. Now, EditorContainer is called as a JSX component
when being rendered.